requirements.txt
numpy
matplotlib
scipy
scikit-image
filterpy
mysql-connector-python
opencv-python
cvzone
lap
torch
ultralytics
datetime